putting marginalized gaussian noise class in a seperate file
